enhancing code organization inside ui package
